AI-Powered-Product-Finder/Product-Finder-Alloydb/Backend/Chatbot-MCP/Chatbot-MCP-Server/requirements.txt
pandas==2.3.3
SQLAlchemy==2.0.45
google-cloud-aiplatform
google-cloud-alloydb-connector[pg8000]==1.11.0
fastmcp==2.14.0
python-dotenv==1.1.1